opt: more optimized split
opt: remove ReuniteError Debug constraint
