+++ monoio/src/net/tcp/mod.rs
@@ -6,7 +6,5 @@ mod split;
 mod stream;
 
 pub use listener::TcpListener;
-pub use split::{
-    ReuniteError as TcpReuniteError, TcpOwnedReadHalf, TcpOwnedWriteHalf, TcpReadHalf, TcpWriteHalf,
-};
+pub use split::{TcpOwnedReadHalf, TcpOwnedWriteHalf, TcpReadHalf, TcpWriteHalf};
 pub use stream::TcpStream;
